Move Audit Into DataFromDatabaseServlet, Delete ReasonServlet and JS Audit
