+++ src/main/java/com/google/sps/servlets/DataFromDatabaseServlet.java
@@ -1,17 +1,22 @@
 package com.google.sps.servlets;
 
+import static com.google.cloud.spanner.TransactionRunner.TransactionCallable;
+
 import com.google.auto.value.AutoValue;
 import com.google.cloud.ByteArray;
 import com.google.cloud.spanner.DatabaseClient;
 import com.google.cloud.spanner.DatabaseId;
+import com.google.cloud.spanner.ReadOnlyTransaction;
 import com.google.cloud.spanner.ResultSet;
 import com.google.cloud.spanner.Spanner;
 import com.google.cloud.spanner.SpannerOptions;
 import com.google.cloud.spanner.Statement;
+import com.google.cloud.spanner.TransactionContext;
 import com.google.common.collect.ImmutableList;
 import com.google.gson.Gson;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -35,6 +40,9 @@ public class DataFromDatabaseServlet extends HttpServlet {
     response.setContentType(constants.TEXT_TYPE);
     selectedTables = request.getParameterValues(constants.TABLE_SELECT_PARAM);
     String databaseName = request.getParameter(constants.DATABASE_PARAM);
+    String reason = request.getParameter("reason");
+    String account = LoginServlet.getCurrentUser();
+    String queryAudit = "";
     initDatabaseClient(databaseName);
 
     List<Table> tables = new ArrayList<>();
@@ -50,6 +58,9 @@ public class DataFromDatabaseServlet extends HttpServlet {
         Table.Builder tableBuilder = Table.builder().setName(table);
         tableBuilder.setColumnSchemas(columnSchemas);
         Statement queryStatement = constructQueryStatement(columnSchemas, table);
+
+        queryAudit += queryStatement + "; ";
+
         executeTableQuery(tableBuilder, queryStatement, columnSchemas);
         
         Table tableObject = tableBuilder.build();
@@ -57,6 +68,7 @@ public class DataFromDatabaseServlet extends HttpServlet {
       }
     }
     String json = new Gson().toJson(tables);
+    insertUsingDml(account,reason,queryAudit);
     response.getWriter().println(json);
   }
 
@@ -181,5 +193,24 @@ public class DataFromDatabaseServlet extends HttpServlet {
     } catch (UnsupportedEncodingException e) {
       return constants.ENCODING_ERROR;
     }
+  }
+
+  private void insertUsingDml(String account, String reason, String query) {
+    this.dbClient = DatabaseConnector.getInstance().getDbClient("example-db");
+    Instant timestamp = Instant.now();
+    String timeString = timestamp.toString();
+    dbClient
+    .readWriteTransaction()
+    .run(
+        new TransactionCallable<Void>() {
+            @Override
+            public Void run(TransactionContext transaction) throws Exception {
+              String sql =
+                  "INSERT INTO AuditLog (Account, Query, Reason, Timestamp) "
+                      + String.format(" VALUES ('%s', '%s', '%s', '%s')",account,query,reason,timeString);
+              transaction.executeUpdate(Statement.of(sql));
+              return null;
+            }
+        });
   }
 }
