+++ src/main/webapp/database-tables.js
@@ -1,50 +1,3 @@
-function audit() {
-    
-    // Create invisible form to send the data
-    var form = document.createElement("form"); 
-    form.setAttribute("type", "hidden");
-    form.setAttribute("id","form");
-    form.setAttribute("method", "post"); 
-    form.setAttribute("action", "/reason"); 
-  
-    // Create account element
-    var account = document.createElement("input"); 
-    account.setAttribute("type", "hidden");
-    account.setAttribute("name", "account");
-    var userVal = document.getElementById("user").innerText;
-    account.setAttribute("value", userVal); 
-
-    const params = new URLSearchParams(window.location.search);
-    var tableSelect = params.getAll('table-select');
-
-    // Create tables accessed  element
-    var queryInput = document.createElement("input"); 
-    queryInput.setAttribute("type", "hidden");
-    queryInput.setAttribute("name", "query"); 
-    queryInput.setAttribute("value", tableSelect);
-
-    // Create reason element
-    var reason = params.get('reason');
-    var reasonInput = document.createElement("input"); 
-    reasonInput.setAttribute("type", "hidden");
-    reasonInput.setAttribute("name", "reason"); 
-    reasonInput.setAttribute("value", reason);
-        
-    // Create queryString element
-    var queryString = window.location.search;
-    var queryStringInput = document.createElement("input"); 
-    queryStringInput.setAttribute("type", "hidden");
-    queryStringInput.setAttribute("name", "queryString"); 
-    queryStringInput.setAttribute("value", queryString);
-
-    // Append everything to form then page
-    form.appendChild(account);  
-    form.appendChild(queryInput);  
-    form.appendChild(queryStringInput);
-    form.appendChild(reasonInput);
-    document.body.appendChild(form); 
-    document.getElementById("form").submit();
-}
 
 //Boolean to ensure data only shows after first click
 var showing = Boolean(false)
