+++ src/main/webapp/main-page.html
@@ -19,7 +19,7 @@
       <div class="button-box">
         <a href="index.html"><button>Splash Page</button></a>
         <button id="filter-button" onmouseover="luckyFilter()" onclick="showFilterPanel()" onmouseout="unlucky()">Show Filters </button>
-        <button id="data-button" onmouseover="luckyData()" onclick="showDatabase(); audit()" onmouseout="unlucky()">Load Data </button>
+        <button id="data-button" onmouseover="luckyData()" onclick="showDatabase();" onmouseout="unlucky()">Load Data </button>
         <button onclick="showShare()" >Share Query</button>
       </div>
     </div>
